Prepare implementation of throttling (to limit rate of requests)
I planned to measure the time of each request and ensure that
no more than the maximum of 8 requests were sent in a floating
window of 1 second. This requires to measure the time of each
request with subsecond accuracy and to maintain a list of the
times measured for up to the maximum number of requests allowed.

However, POSIX utilities only offer an accuracy of 1 second,
which requires to stretch the floating window to 2 seconds,
or to halve the allowed rate to 4 requests per second. These
two methods will result in similar request rates in average,
but the latter is simpler to implement and reduces the maximum
request rate observed in a floating window.

I will thus simplify the logic on these lines:
* count the number of requests in current second
* reset the counter when the second changes