Remove TODO comment, checked, not needed
Although the latest date is in the future, the data is actually
available because it corresponds to sales in the past, and already
available before the date of release.

In this case, although I intended to only download best seller lists
if the file is not already present, it may be interesting to refresh
the file unconditionally to check that it has not been modified in
the meantime.