Complete step 3: create a separate source for each best seller list
For each date of publication (weekly), a separate source is created
in its own folder, acquired and parsed to extract CSV data with the
list of best selling manga for a week.

I have also added a log utility to print messages both to the standard
output and the file log.txt. This utility is used outside the loop;
the whole output of the loop is redirected in a similar manner to
both standard output and log.txt.