Add step 3: crosscheck
I have simply ordered manga titles and authors from the NYT's best sellers,
which makes it clear that the author names and series titles are not
normalized: they vary from volume to volume within a single series.

Also, the primary ISBN10 and ISBN13 do not correspond to the first volume
of the series, as I expected, but to the current volume; the link
between volumes of a single series is not formalized in these lists.

I will complete this step with additional cross-references
once I have added a second source of manga information
from MangaConseil.com.